refactor(transform): move ast_rewrite to shared/ and migrate the server template-rune pass
The collect-and-splice toolkit is not client-specific — relocate
`client/ast_rewrite.rs` to `shared/ast_rewrite.rs` (re-exported as
`client::ast_rewrite` so the 26 client passes are untouched) and route the
last hand-rolled server pass, `server::template_rune_ast`, through it.

`transform_template_rune_ast`'s previously unbounded nesting loop now shares
the `MAX_FIXED_POINT_ITERS`-bounded `fixed_point` driver (identical output for
all converging inputs, no infinite-loop risk), and its parse/splice scaffold
collapses to `rewrite_once`.

With this, no `*_ast.rs` pass in Phase 3 — client or server — hand-rolls
arena take/restore, parse-error bail, edit splicing, or the fixed-point loop;
they all go through one shared, unit-tested toolkit.

Byte-identical: lib tests (1283) and compiler_fixtures / runtime / ssr all green.
