+++ crates/rsvelte_core/src/compiler/phases/3_transform/shared/ast_rewrite.rs
