+++ crates/rsvelte_core/src/compiler/phases/3_transform/shared/mod.rs
@@ -4,6 +4,7 @@
 //! code generation.
 
 pub mod assignments;
+pub mod ast_rewrite;
 pub mod async_body;
 pub mod template;
